Adding more improvements

+++ src/main/java/com/ninjaone/dundie_awards/repository/DundieAwardRepository.java
@@ -1,9 +1,9 @@
 package com.ninjaone.dundie_awards.repository;
 
-import com.ninjaone.dundie_awards.domain.DundieAward;
+import com.ninjaone.dundie_awards.model.DundieAward;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DundieAwardRepository extends JpaRepository<DundieAward, Long> {
+public interface DundieAwardRepository extends JpaRepository<DundieAward, Integer> {
 }
